Modifica da parseInt a ParseFloat nei km

+++ script.js
@@ -1,14 +1,13 @@
 var km = prompt("inserire il numero di km percorsi:");
 var eta = prompt("inserire la propria eta':");
 
-parseInt(km);
+parseFloat(km);
 parseInt(eta);
 
 var costo;
 parseFloat(costo);
 
 if (eta<18){
-
     costo = ((km*0.21)-((km*0.21)/100*20));
     console.log("Il costo del biglietto è uguale a: " + costo);
 } else if (eta>65){
